+++ ecm1240.go
@@ -8,12 +8,6 @@ import (
 	"github.com/wangdrew/powermonitor-go/models"
 	"time"
 )
-
-type Source interface {
-	Init() error
-	Read() (models.PowerMetrics, error)
-}
-
 func NewECM1240Source(deviceName, serialPath string) Source {
 	return &ECM1240Source{Name: deviceName, serialPath: serialPath}
 }
